Fix: created a new file for a function in game_loop.c
